Changed grid to run, removed original run file
